LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-alignment-implies-size-change-031-expected.txt



FAIL .before 1 assert_equals:
<img id="item" src="support/100x100-green.png" class=" before" data-expected-width="200" data-expected-height="200">
height expected 200 but got 100
PASS .after 2
